+++ packages/components-react/accordion-react/src/noSideEffects.ts
