+++ packages/components-react/alert-react/src/noSideEffects.ts
