+++ packages/components-react/alert-react/vite.config.ts
@@ -8,7 +8,6 @@ export default defineConfig({
       // Could also be a dictionary or array of multiple entry points
       entry: {
         index: resolve(__dirname, 'src/index.ts'),
-        noCss: resolve(__dirname, 'src/noCss.ts'),
       },
       name: 'Alert',
       // the proper extensions will be added
