+++ packages/components-react/blockquote-react/src/noSideEffects.ts
