+++ packages/components-react/breadcrumb-nav-react/src/noSideEffects.ts
