+++ packages/components-react/card-as-link-react/src/noSideEffects.ts
