+++ packages/components-react/dot-badge-react/src/noSideEffects.ts
