+++ packages/components-react/figure-react/src/noSideEffects.ts
