+++ packages/components-react/icon-react/src/noSideEffects.ts
