+++ packages/components-react/image-react/src/noSideEffects.ts
