+++ packages/components-react/library-react/package.json
@@ -39,10 +39,10 @@
       "import": "./dist/index.mjs",
       "require": "./dist/index.cjs.js"
     },
-    "./no-css": {
-      "rhc:source": "./src/noCss.ts",
-      "types": "./dist/noCss.d.ts",
-      "import": "./dist/noCss.mjs"
+    "./no-side-effects": {
+      "rhc:source": "./src/noSideEffects.ts",
+      "types": "./dist/noSideEffects.d.ts",
+      "import": "./dist/noSideEffects.mjs"
     },
     "./CHANGELOG.md": "./CHANGELOG.md",
     "./README.md": "./README.md"
