+++ packages/components-react/library-react/src/index.ts
@@ -1,3 +1,3 @@
 import '@rijkshuisstijl-community/components-css/dist/index.css';
 
-export * from './noCss';
+export * from './noSideEffects';
