+++ packages/components-react/link-react/src/noSideEffects.ts
