+++ packages/components-react/paragraph-react/src/noSideEffects.ts
