+++ packages/components-react/separator-react/package.json
@@ -19,10 +19,10 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.mjs"
     },
-    "./no-css": {
-      "rhc:source": "./src/noCss.ts",
-      "types": "./dist/noCss.d.ts",
-      "import": "./dist/noCss.mjs"
+    "./no-side-effects": {
+      "rhc:source": "./src/noSideEffects.ts",
+      "types": "./dist/noSideEffects.d.ts",
+      "import": "./dist/noSideEffects.mjs"
     }
   },
   "scripts": {
